http://jira.aplana.com/browse/SBRFACCTAX-7516 : Реализация флага Действующего подразделения.

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/refbook/RefBookDepartmentDao.java
@@ -78,7 +78,7 @@ public interface RefBookDepartmentDao {
      */
     public List<Long> getPeriodsByTaxTypesAndDepartments(List<TaxType> taxTypes, List<Integer> departmentList);
 
-    void update(long uniqueId, Map<String, RefBookValue> records, List<RefBookAttribute> attributes);
+    void update(int uniqueId, Map<String, RefBookValue> records, List<RefBookAttribute> attributes);
 
     int create(Map<String, RefBookValue> records, List<RefBookAttribute> attributes);
